+++ core/src/main/java/com/bc/fiduceo/reader/slstr_subset/SlstrRegriddedSubsetReaderPluginOblique.java
@@ -1,0 +1,23 @@
+package com.bc.fiduceo.reader.slstr_subset;
+
+import com.bc.fiduceo.reader.DataType;
+import com.bc.fiduceo.reader.Reader;
+import com.bc.fiduceo.reader.ReaderContext;
+import com.bc.fiduceo.reader.ReaderPlugin;
+
+public class SlstrRegriddedSubsetReaderPluginOblique implements ReaderPlugin {
+    @Override
+    public String[] getSupportedSensorKeys() {
+        return new String[]{"slstr-s3a-uor-o", "slstr-s3b-uor-o"};
+    }
+
+    @Override
+    public Reader createReader(ReaderContext readerContext) {
+        return new SlstrRegriddedSubsetReader(readerContext, false);
+    }
+
+    @Override
+    public DataType getDataType() {
+        return DataType.POLAR_ORBITING_SATELLITE;
+    }
+}