+++ core/src/test/java/com/bc/fiduceo/util/NetCDFUtilsTest.java
@@ -201,6 +201,31 @@ public class NetCDFUtilsTest {
     }
 
     @Test
+    public void testGetGlobalAttributeDouble() {
+        final NetcdfFile netcdfFile = mock(NetcdfFile.class);
+        final Attribute attribute = mock(Attribute.class);
+
+        when(attribute.getNumericValue()).thenReturn(23.56);
+        when(netcdfFile.findGlobalAttribute("the_attribute")).thenReturn(attribute);
+
+        final double doubleValue = NetCDFUtils.getGlobalAttributeDouble("the_attribute", netcdfFile);
+        assertEquals(23.56, doubleValue, 1e-8);
+    }
+
+    @Test
+    public void testGetGlobalAttributeDouble_missingAttribute() {
+        final NetcdfFile netcdfFile = mock(NetcdfFile.class);
+
+        when(netcdfFile.findGlobalAttribute("the_attribute")).thenReturn(null);
+
+        try {
+            NetCDFUtils.getGlobalAttributeDouble("the_attribute", netcdfFile);
+            fail("RuntimeException expected");
+        } catch (RuntimeException expected) {
+        }
+    }
+
+    @Test
     public void testGetFillValue_fromAttribute() {
         final Variable variable = mock(Variable.class);
         final Attribute attribute = mock(Attribute.class);